pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Irina Ivashko <ira.ivashko.99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.2"
pyjwt = "^2.8.0"
fastapi = "^0.111.1"
uvicorn = "^0.30.5"
aiokafka = "^0.11.0"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.32"
psycopg2 = "^2.9.9"
bcrypt = "^4.2.0"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.19.2"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
